requirements.txt
Django==1.5
Fabric==1.6.0
Pillow==2.5.3
South==0.7.6
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
balanced==1.1.1
billiard==3.3.0.18
celery==3.1.8
certifi==0.0.8
chardet==2.2.1
django-celery==3.1.1
django-extensions==1.1.1
django-tastypie==0.9.14
iso8601==0.1.4
kombu==3.0.23
mimeparse==0.1.3
mock==0.8.0
paramiko==1.10.1
pycrypto==2.6
python-dateutil==2.1
pytz==2013b
requests==2.4.1
simplejson==3.6.3
six==1.2.0
uritemplate==0.6
wac==0.23
wsgiref==0.1.2
xlrd==0.9.3

